move admin subscription route to admin folder and add route to get users subscription plan
